telegram_bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Anton <109826223+Arcadiyyyyyyyy@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-telegram-bot = "^20.6"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
